+++ .gitignore
@@ -1,0 +1,2 @@
+
+ReadBi_0_01/APIConstants.swift
